requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile 
#
click-config-file==0.5.0
click==7.0
colorama==0.4.1
configobj==5.0.6          # via click-config-file
pyyaml==5.1
six==1.12.0               # via configobj
